scripts: drop nuxtron
Since we're not actually using nuxtron for dev or build, drop it and
refactor the scripts to use a common script.
