+++ package.json
@@ -2,10 +2,10 @@
   "name": "rancher-desktop",
   "version": "0.0.1",
   "scripts": {
-    "dev": "node ./scripts/dev.mjs",
+    "dev": "node ./scripts/build.mjs dev",
     "lint": "eslint --ignore-path=.gitignore --ext mjs,js,vue --fix .",
     "lint:nofix": "eslint --ignore-path=.gitignore --ext mjs,js,vue .",
-    "build": "node ./scripts/build.mjs",
+    "build": "node ./scripts/build.mjs build",
     "test": "npm run lint:nofix && npm run test:unit",
     "test:unit": "jest",
     "test:unit:watch": "npm run test:unit -- --watch",
@@ -57,7 +57,6 @@
     "follow-redirects": "^1.13.1",
     "node-sass": "^5.0.0",
     "nuxt": "^2.14.12",
-    "nuxtron": "^0.3.1",
     "sass-loader": "^10.1.1",
     "vue-jest": "^3.0.7",
     "vue-template-compiler": "^2.6.12",
